Module google_iamcredentials1::api

Structs

GenerateAccessTokenRequest
There is no detailed description.

GenerateAccessTokenResponse
There is no detailed description.

GenerateIdTokenRequest
There is no detailed description.

GenerateIdTokenResponse
There is no detailed description.

IAMCredentials
Central instance to access all IAMCredentials related resource activities

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the IAMCredentials hub.

ProjectServiceAccountGenerateAccessTokenCall
Generates an OAuth 2.0 access token for a service account.

ProjectServiceAccountGenerateIdTokenCall
Generates an OpenID Connect ID token for a service account.

ProjectServiceAccountSignBlobCall
Signs a blob using a service account's system-managed private key.

ProjectServiceAccountSignJwtCall
Signs a JWT using a service account's system-managed private key.

SignBlobRequest
There is no detailed description.

SignBlobResponse
There is no detailed description.

SignJwtRequest
There is no detailed description.

SignJwtResponse
There is no detailed description.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.